Fix code 128: fatal: detected dubious ownership in repository
This happens the second time we try to use ansible git module to
clone a repo. It complains the ownership is wrong, cause we
cloned it using root user and using chmod to change it back
at the end of the main.yml.

Solution based on https://community.atlassian.com/t5/Bitbucket-questions/Getting-code-128-fatal-detected-dubious-ownership-in-repository/qaq-p/2373506

Signed-off-by: tarbeleliyahu <tarbeleliyahu@paypal.com>